Move vec2 utils
