+++ web_ui/packages/smart-tools/src/utils/vec2.test.ts
