+++ web_ui/src/pages/annotator/tools/edit-tool/edit-keypoint/rotation-anchor.component.tsx
@@ -1,12 +1,12 @@
 // Copyright (C) 2022-2025 Intel Corporation
 // LIMITED EDGE SOFTWARE DISTRIBUTION LICENSE
 
+import { Vec2 } from '@geti/smart-tools';
 import { Rotation } from '@geti/ui/icons';
 import clsx from 'clsx';
 
 import { radiansToDegrees } from '../../../../../core/annotations/math';
 import { Point } from '../../../../../core/annotations/shapes.interface';
-import * as Vec2 from '../../../../../core/annotations/vec2';
 import { Anchor } from '../anchor.component';
 
 import styles from './rotation-anchor.module.scss';
