+++ web_ui/src/pages/annotator/tools/edit-tool/edit-rotated-bounding-box/location.ts
@@ -1,9 +1,10 @@
 // Copyright (C) 2022-2025 Intel Corporation
 // LIMITED EDGE SOFTWARE DISTRIBUTION LICENSE
 
+import { Vec2 } from '@geti/smart-tools';
+
 import { transformPointInRotatedRectToScreenSpace } from '../../../../../core/annotations/rotated-rect-math';
 import { Point, RotatedRect } from '../../../../../core/annotations/shapes.interface';
-import * as Vec2 from '../../../../../core/annotations/vec2';
 
 const getLocations = (shape: RotatedRect, gap: number) => {
     const top = { x: shape.x, y: shape.y - shape.height / 2 };
