+++ web_ui/src/pages/annotator/tools/edit-tool/edit-rotated-bounding-box/rotation-element/rotation-element.component.tsx
@@ -3,11 +3,12 @@
 
 import { useEffect, useState } from 'react';
 
+import { Vec2 } from '@geti/smart-tools';
+
 import { RegionOfInterest } from '../../../../../../core/annotations/annotation.interface';
 import { radiansToDegrees } from '../../../../../../core/annotations/math';
 import { RotatedRect } from '../../../../../../core/annotations/shapes.interface';
 import { ShapeType } from '../../../../../../core/annotations/shapetype.enum';
-import * as Vec2 from '../../../../../../core/annotations/vec2';
 import { Line } from '../../../../annotation/shapes/line.component';
 import { SideAnchorLocationsProps } from '../location';
 import { RotationAnchor } from './rotation-anchor.component';
